organize di modules in chat data module
