+++ core_chat_data/src/commonMain/kotlin/net/thechance/mena/core_chat/data/contacts/ContactMapper.kt
@@ -1,7 +1,7 @@
 package net.thechance.mena.core_chat.data.contacts
 
-import net.thechance.mena.core_chat.data.contacts.source.dto.ContactCreationRequestDto
-import net.thechance.mena.core_chat.data.contacts.source.dto.ContactDto
+import net.thechance.mena.core_chat.data.contacts.dto.ContactCreationRequestDto
+import net.thechance.mena.core_chat.data.contacts.dto.ContactDto
 import net.thechance.mena.core_chat.data.shared.dto.PagedDataDto
 import net.thechance.mena.core_chat.domain.entity.Contact
 import net.thechance.mena.core_chat.domain.exception.ContactsFetchFailedException
