+++ core_chat_data/src/commonMain/kotlin/net/thechance/mena/core_chat/data/contacts/dto/ContactDto.kt
@@ -1,4 +1,4 @@
-package net.thechance.mena.core_chat.data.contacts.source.dto
+package net.thechance.mena.core_chat.data.contacts.dto
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
