+++ core_chat_data/src/commonMain/kotlin/net/thechance/mena/core_chat/data/di/dataProviderModule.kt
@@ -1,0 +1,11 @@
+package net.thechance.mena.core_chat.data.di
+
+import com.bilalazzam.contacts_provider.ContactsProvider
+import org.koin.core.scope.Scope
+import org.koin.dsl.module
+
+internal val dataProviderModule = module {
+    single { createContactsProvider() }
+}
+
+expect fun Scope.createContactsProvider(): ContactsProvider
