+++ core_chat_data/src/commonMain/kotlin/net/thechance/mena/core_chat/data/di/networkModule.kt
@@ -1,6 +1,5 @@
 package net.thechance.mena.core_chat.data.di
 
-import com.bilalazzam.contacts_provider.ContactsProvider
 import io.ktor.client.engine.HttpClientEngineConfig
 import io.ktor.client.engine.HttpClientEngineFactory
 import net.thechance.mena.core_chat.data.network.createHttpClient
@@ -8,12 +7,10 @@ import org.koin.core.qualifier.named
 import org.koin.core.scope.Scope
 import org.koin.dsl.module
 
-internal val serviceModule = module {
-    single(named("baseUrl")) { "http://10.0.2.2:8080" }
-    single { createContactsProvider() }
+internal val networkModule = module {
+    single(named("baseUrl")) { "http://10.0.2.2:8080" } // TODO change to real url in production
     single { createHttpClientEngine() }
     single { createHttpClient(get(named("baseUrl"))) }
 }
 
-expect fun Scope.createContactsProvider(): ContactsProvider
 expect fun Scope.createHttpClientEngine(): HttpClientEngineFactory<HttpClientEngineConfig>
